fixed warnings

+++ include/render/rtshadowpass.h
@@ -69,6 +69,6 @@ public:
     void setLightsBuffer(VkBuffer buffer);
     void setGbuffer(GBuffer* gbuffer);
     void setOutputImageView(VkImageView imageView);
-    void updateUniformBuffer(uint32_t currentImage, uint32_t frameNumber, Scene& scene, uint32_t cameraIndex, const uint32_t width, const uint32_t height);
+    void updateUniformBuffer(uint32_t currentImage, uint64_t frameNumber, Scene& scene, uint32_t cameraIndex, const uint32_t width, const uint32_t height);
 };
 } // namespace nevk
